refactor(drift_sync): extract context.py (2/8)
Move ScanContext + DriftError + DriftOutputError out of __init__.py
into a new context.py submodule. __init__.py re-exports all three for
backward compatibility.

Unused imports after extraction (dataclass, BranchProtectionConfig,
LabelsConfig, ProfileSpec) removed from __init__.py.

Refs #47 (Theme A item 4).
